backend/app/requirements.txt
flask
flask-cors
requests
google-generativeai
python-dotenv
gitpython
networkx
time